+++ hummingbird/ml/containers/sklearn/__init__.py
@@ -1,0 +1,9 @@
+# -------------------------------------------------------------------------
+# Copyright (c) Microsoft Corporation. All rights reserved.
+# Licensed under the MIT License. See License.txt in the project root for
+# license information.
+# --------------------------------------------------------------------------
+
+"""
+Output containers matching the Scikit-learn API are listed under this package.
+"""
